requirements.txt
amqp==1.4.6
anyjson==0.3.3
billiard==3.3.0.20
boto==2.38.0
celery==3.1.18
cffi==1.2.1
coverage==4.0.2
cryptography==1.0.1
Django==1.9.4
django-allauth==0.23.0
django-boto==0.3.9
django-braces==1.8.1
django-celery==3.1.16
django-contrib-comments==1.6.1
django-dbbackup==2.0.4
django-debug-toolbar==1.3.2
django-easy-timezones==0.7.0
django-oauth-toolkit==0.9.0
django-redis-cache==1.6.0
django-rest-framework-social-oauth2==1.0.2
djangorestframework==3.2.3
ecdsa==0.13
enum34==1.0.4
gunicorn==19.3.0
idna==2.0
ipaddress==1.0.14
kombu==3.0.26
ndg-httpsclient==0.4.0
oauthlib==1.0.1
paramiko==1.15.2
Pillow==3.1.1
psycopg2==2.6.1
pyasn1==0.1.8
pycparser==2.14
pycrypto==2.6.1
pygeoip==0.3.2
PyJWT==1.4.0
pyOpenSSL==0.15.1
pysftp==0.2.8
python-dateutil==2.4.2
python-openid==2.2.5
python-social-auth==0.2.12
pytz==2015.7
redis==2.10.3
requests==2.7.0
requests-oauthlib==0.5.0
six==1.9.0
sqlparse==0.1.16
